backend/requirements.txt
pandas
numpy
scikit-learn
openpyxl
dotenv
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
python-multipart
openai
passlib[bcrypt]
python-jose[cryptography]
requests